requirements.txt
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf
pytest
python-dotenv
tiktoken
streamlit
